fix: move to domain package
